Pre-extract Station data class to commonMain
Plan correction: the build-time generator emits TubeStationsData.kt
into commonMain referencing io.kotest.property.arbs.tube.Station.
That blocked compileKotlinJs because Station lived in jvmMain. The
pure data class has no platform dependencies, so move just it now;
T8 still owns the rest of tube.kt (java.time → kotlinx-datetime,
arb functions, Journey/FareMethod).

After this commit compileKotlinJvm and compileKotlinJs both pass.
