+++ src/jvmMain/kotlin/io/kotest/property/arbs/tube/tube.kt
@@ -10,16 +10,6 @@ import io.kotest.property.arbs.loadResource
 import java.time.LocalDateTime
 import kotlin.random.Random
 
-data class Station(
-  val id: Long,
-  val latitude: Double,
-  val longitude: Double,
-  val name: String,
-  val zone: Double,
-  val lines: Int,
-  val rail: Int
-)
-
 enum class FareMethod {
   Oyster, Contactless, NetworkRail, Mobile
 }
